Reorganize how docstring is generated/incorporated
The formatting of the docstring has been outsourced to a separate
function in formatting.add_attribute_section(). This way, the
render_docstrings() only needs to call that function and pass its
return to the __doc__ attribute of the base class (and subclasses).

There are more suggestions in the PR but I had the impression that it's
beneficial to update the recent code.