Include a test for `add_attribute_section()`
With this commit, a test for the autodoc-feature will be implemented.
For now, however, pytest returns an AssertionError because
`add_attribute_section()` in `test_add_attribute_section()` returns
the docstring twice. Reason unknown.

Also, `@pytest.mark.xfail` was introduced to circumvent failures in
`test_process_decorator()` and `test_add_attribute_section()`. This
will be only temporary.